Break types into their own files.
